feat: 결제요청 api 두 번 호출 방지를 위한 상위 컴포넌트에서 호출 방식 변경
